+++ store/src/main/java/com/zimbra/cs/db/DbSearch.java
@@ -996,15 +996,15 @@ public final class DbSearch {
       case PRIORITY:
         return Strings.nullToEmpty(rs.getString(SORT_COLUMN_ALIAS));
       case SIZE:
-        return Long.valueOf(rs.getInt(SORT_COLUMN_ALIAS));
+        return (long) rs.getInt(SORT_COLUMN_ALIAS);
       case ID:
-        return Integer.valueOf(rs.getInt(SORT_COLUMN_ALIAS));
+        return rs.getInt(SORT_COLUMN_ALIAS);
       case DATE:
       default:
         // Assuming this multiplication by 1000 is intended for DATE in order to convert from a
         // UNIX time in seconds to milliseconds since epoc
         // seems odd to also do this for the default case though...
-        return Long.valueOf(rs.getInt(SORT_COLUMN_ALIAS) * 1000L);
+        return rs.getInt(SORT_COLUMN_ALIAS) * 1000L;
     }
   }
